offline_package/requirements_raspi.txt
# 라즈베리파이4용 Python 패키지 requirements
# MediaPipe와 OpenCV를 라즈베리파이에 최적화된 버전으로 설치

# 핵심 컴퓨터 비전
opencv-python==4.8.1.78
mediapipe==0.10.3

# 게임 엔진
pygame==2.5.2

# 이미지 처리
Pillow==10.0.1
numpy==1.24.3

# 오디오 처리
pydub==0.25.1

# 시스템 유틸리티
psutil==5.9.5

# JSON 처리 (내장 모듈이지만 명시)
# json - 내장 모듈

# 기타 유틸리티
pathlib2==2.3.7
